#2 Fixed failures about not being able to get the pods
First, I was getting an error that I was missing Service Account,even though it was there, turns out, I had specified nsand the other resources had no ns.

After fixing that, I ran into CrashLoopBackOff, and it turns out, the issue was with my image - the ENTRYPOINT was no found since there was an error copying the executable from the builder so it was just failing.

Now that is fixed, next stop is ome bit of clean-up and playing around with the NetworkPolicy